+++ allauth/account/internal/flows/password_reset.py
@@ -1,15 +1,20 @@
+from typing import Optional
 from urllib.parse import quote
 
 from django.contrib import messages
 from django.contrib.auth.models import AbstractBaseUser
-from django.http import HttpRequest
+from django.http import HttpRequest, HttpResponse
 from django.urls import reverse
 
 from allauth.account import app_settings, signals
 from allauth.account.adapter import get_adapter
 from allauth.account.app_settings import LoginMethod
+from allauth.account.internal.flows.login import (
+    perform_login,
+    record_authentication,
+)
 from allauth.account.internal.flows.signup import send_unknown_account_mail
-from allauth.account.models import EmailAddress
+from allauth.account.models import EmailAddress, Login
 from allauth.core.internal.httpkit import get_frontend_url
 from allauth.utils import build_absolute_uri
 
@@ -18,13 +23,31 @@ def reset_password(user: AbstractBaseUser, password: str) -> None:
     get_adapter().set_password(user, password)
 
 
-def finalize_password_reset(request: HttpRequest, user: AbstractBaseUser) -> None:
+def perform_password_reset_login(
+    request: HttpRequest,
+    user: AbstractBaseUser,
+    phone: Optional[str] = None,
+    email: Optional[str] = None,
+) -> HttpResponse:
+    extra_data = {}
+    if phone:
+        extra_data["phone"] = phone
+    elif email:
+        extra_data["email"] = email
+    record_authentication(request, method="password_reset", **extra_data)
+    login = Login(user=user, email=email)
+    return perform_login(request, login)
+
+
+def finalize_password_reset(
+    request: HttpRequest, user: AbstractBaseUser, email: Optional[str] = None
+) -> Optional[HttpResponse]:
     adapter = get_adapter()
     if user:
         # User successfully reset the password, clear any
         # possible cache entries for all email addresses.
-        for email in EmailAddress.objects.filter(user_id=user.pk):
-            adapter._delete_login_attempts_cached_email(request, email=email.email)
+        for address in EmailAddress.objects.filter(user_id=user.pk):
+            adapter._delete_login_attempts_cached_email(request, email=address.email)
 
     adapter.add_message(
         request,
@@ -37,6 +60,9 @@ def finalize_password_reset(request: HttpRequest, user: AbstractBaseUser) -> Non
         user=user,
     )
     adapter.send_notification_mail("account/email/password_reset", user)
+    if app_settings.LOGIN_ON_PASSWORD_RESET:
+        return perform_password_reset_login(request, user, email=email)
+    return None
 
 
 def get_reset_password_url(request: HttpRequest) -> str: